% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JRNMM_Splitting_Cpp.R
\name{covariance_matrix_JRNMM}
\alias{covariance_matrix_JRNMM}
\title{covariance_matrix_JRNMM}
\usage{
covariance_matrix_JRNMM(dGamma, dSigma, h)
}
\arguments{
\item{dGamma}{vector of diagonal entries of the Gamma matrix}

\item{dSigma}{vector of diagonal entries of the Sigma matrix}

\item{h}{step size used for path simulation}
}
\value{
See description
}
\description{
Covariance matrix of the linear SDE in the splitting scheme
}
