% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JRNMM_Splitting_Cpp.R
\name{exponential_matrix_JRNMM}
\alias{exponential_matrix_JRNMM}
\title{exponential_matrix_JRNMM}
\usage{
exponential_matrix_JRNMM(dGamma, h)
}
\arguments{
\item{dGamma}{vector of diagonal entries of the Gamma matrix}

\item{h}{step size used for path simulation}
}
\value{
path of the stochastic N-population JRNMM
}
\description{
Calculation of the diagonal entering into the block matrix of
}
