% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JRNMM_Splitting_Cpp.R
\name{fast_JRNMM_Splitting}
\alias{fast_JRNMM_Splitting}
\title{fast_JRNMM_Splitting}
\usage{
fast_JRNMM_Splitting(N, grid, h, startv, dGamma, dSigma, Theta, Rho, K)
}
\arguments{
\item{N}{number of populations of neurons}

\item{grid}{time points at which to simulate the process}

\item{h}{step size used for path simulation}

\item{startv}{starting point x0}

\item{dGamma}{vector of diagonal entries of the Gamma matrix}

\item{dSigma}{vector of diagonal entries of the Sigma matrix}

\item{Theta}{vector of continuous parameters}

\item{Rho}{matrix of {0,1} discrete parameters}

\item{K}{Matrix of strength parameters
#'@return path of the stochastic N-population JRNMM}
}
\description{
(faster version of the) Strang splitting method for path simulation of the
stochastic multi-population JRNMM
}
