% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JRNMM_Splitting_Cpp.R
\name{JRNMM_Splitting}
\alias{JRNMM_Splitting}
\title{JRNMM_Splitting}
\usage{
JRNMM_Splitting(N, grid, h, startv, dm, meanVec, covMat, Theta, Rho, K)
}
\arguments{
\item{N}{number of populations of neurons}

\item{grid}{time points at which to simulate the process}

\item{h}{step size used for path simulation}

\item{startv}{starting point x0}

\item{dm}{exponential matrix coming from the linear SDE in the splitting scheme}

\item{meanVec}{6N dimensional vector of 0s}

\item{covMat}{covariance matrix coming from the linear SDE in the splitting scheme}

\item{Theta}{vector of continuous parameters}

\item{Rho}{matrix of {0,1} discrete parameters}

\item{K}{Matrix of strength parameters
#'@return path of the stochastic N-population JRNMM}
}
\description{
Strang splitting method for path simulation of the
stochastic multi-population JRNMM
}
