% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JRNMM_Splitting_Cpp.R
\name{KmatrixgivenLc}
\alias{KmatrixgivenLc}
\title{KmatrixgivenLc}
\usage{
KmatrixgivenLc(N, L, c)
}
\arguments{
\item{N}{number of populations of neurons}

\item{L}{strength parameter}

\item{c}{[0,1] parameter}
}
\value{
Kmatrix given L and c
}
\description{
Code to calculate the K matrix given L and c
}
