% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JRNMM_Splitting_Cpp.R
\name{linear_JRNMM}
\alias{linear_JRNMM}
\title{linear_JRNMM}
\usage{
linear_JRNMM(vec, dm, xi)
}
\arguments{
\item{vec}{vector coming from the non-linear ODE from the Strang splitting scheme}

\item{dm}{exponential matrix coming from the linear SDE in the splitting scheme}

\item{xi}{gaussian increments}
}
\value{
path of the stochastic N-population JRNMM
}
\description{
linear SDE: exact solution of 1st subequation of splitting procedure for the JRNMM
}
