% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JRNMM_Splitting_Cpp.R
\name{mv_mult_JRNMM}
\alias{mv_mult_JRNMM}
\title{mv_mult_JRNMM}
\usage{
mv_mult_JRNMM(mat, vet)
}
\arguments{
\item{mat}{matrix}

\item{vet}{vector}
}
\value{
See description
}
\description{
Matrix-Vector Multiplication
}
