% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JRNMM_Splitting_Cpp.R
\name{nonlinear_JRNMM}
\alias{nonlinear_JRNMM}
\title{nonlinear_JRNMM}
\usage{
nonlinear_JRNMM(N, vec, h, Theta, Rho, K)
}
\arguments{
\item{N}{number of populations of neurons}

\item{vec}{starting point before applying the solution to the non-linear ODE from the Strang splitting scheme}

\item{h}{step size used for path simulation}

\item{Theta}{vector of continuous parameters}

\item{Rho}{matrix of {0,1} discrete parameters}

\item{K}{Matrix of strength parameters}
}
\value{
See description
}
\description{
nonlinear ODE: exact solution of 2nd subequation of splitting procedure for the JRNMM
}
