% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JRNMM_Splitting_Cpp.R
\name{observedJRNMM}
\alias{observedJRNMM}
\title{observedJRNMM}
\usage{
observedJRNMM(N, grid, h, startv, dGamma, dSigma, Theta, Rho, K)
}
\arguments{
\item{N}{number of populations of neurons}

\item{grid}{time points at which to simulate the process}

\item{h}{step size used for path simulation}

\item{startv}{starting point x0}

\item{dGamma}{vector of diagonal entries of the Gamma matrix}

\item{dSigma}{vector of diagonal entries of the Sigma matrix}

\item{Theta}{vector of continuous parameters}

\item{Rho}{matrix of {0,1} discrete parameters}

\item{K}{Matrix of strength parameters}
}
\value{
observed components of the stochastic N-population JRNMM
}
\description{
Strang splitting method for path simulation of the observed components of the
stochastic multi-population JRNMM.
}
