% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JRNMM_Splitting_Cpp.R
\name{sigmoid_JRNMM}
\alias{sigmoid_JRNMM}
\title{sigmoid_JRNMM}
\usage{
sigmoid_JRNMM(x, vmax, v0, r)
}
\arguments{
\item{x}{value at which the sigmoid function should be evaluated}

\item{vmax}{value of the JRNMM}

\item{v0}{value of the JRNMM}

\item{r}{value of the JRNMM}
}
\value{
path of the stochastic N-population JRNMM
}
\description{
Sigmoid function of the JRNMM
}
