% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StrangSplittingJRNMM.r
\docType{package}
\name{StrangSplittingJRNMM}
\alias{StrangSplittingJRNMM-package}
\alias{StrangSplittingJRNMM}
\title{StrangSplittingJRNMM}
\description{
Simulation of N stochastics JR-NMMs with the Strang splitting scheme
}
\author{
Massimiliano Tamborrino
}
